pyproject.toml
[project]
name = "django-drf-multi-tenancy"
version = "0.1.0"
description = ""
authors = [
    {name = "lmbelo",email = "lucas.belo@live.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==4.0)",
    "djangorestframework (==3.14.0)",
    "django-tenants (==3.5.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "drf-yasg (>=1.21.9,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
